% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefPow.R
\name{coefPow}
\alias{coefPow}
\title{Coefficients and powers}
\usage{
coefPow(x)
}
\arguments{
\item{x}{A vector of powers}
}
\description{
Determines the coefficients and powers needed to express
\deqn{\sum p_{i_1}^{\alpha_1} p_{i_2}^{\alpha_2} \cdots p_{i_n}^{\alpha_n}} in terms of \eqn{\sum p^{m}}.
}
\examples{
# Computing \\sum p_{i}^2 p_{j}^2 for i \\neq j and p = c(.1,.2,.3,.4):
coefPow(c(2,2))
p = c(.1,.2,.3,.4)
-sum(p^4) + sum(p^2)^2

}
\author{
Poul Svante Eriksen, \email{svante@math.aau.dk}
}
