% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dk_str}
\alias{dk_str}
\title{Allele Frequencies of Danish STR Markers}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 72 rows and 11 columns.
}
\usage{
dk_str
}
\description{
Allele Frequencies of Danish STR Markers
}
\keyword{datasets}
