% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA_to_df.R
\name{pack_pedigree_information}
\alias{pack_pedigree_information}
\title{Helper function that packs the information of several pedigrees as input to \code{pedNoA_to_df}}
\usage{
pack_pedigree_information(pedigrees, ids, is_unrelated, markers,
  convoluted = FALSE)
}
\arguments{
\item{pedigrees}{A list with named pedigrees with class \code{linkdat}; see \code{paramlink}.}

\item{ids}{A list of vectors of indicies corresponding to \code{pedigrees}.}

\item{is_unrelated}{A list of the number of unrelated individuals corresponding to \code{pedigrees}.}

\item{markers}{A named list of marker frequencies.}

\item{convoluted}{A boolean indicating if markers should be convoluted}
}
\description{
Calculates the probabilities of sharing \code{0,1,2,...} alleles between \code{ids} and \code{nindep} for the markers in \code{afreq} over several markers
}
\examples{
# To come:

}
