% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA_.R
\name{pedNoA_}
\alias{pedNoA_}
\title{Number of shared alleles}
\usage{
pedNoA_(relation, afreq, verbose = T)
}
\arguments{
\item{relation}{A valid relation; see notes.}

\item{afreq}{A vector of allele frequencies.}

\item{verbose}{If TRUE, information about \code{relation} will be printet.}
}
\description{
Calculates the probabilities of sharing \code{0,1,2,...} alleles between members with the relation \code{relation}
}
\note{
Valid relations are:
"nsibs" with \deqn{n \geq 1} for \code{n} siblings.
"uncle_nephew" for uncle and nephew
"first_cousins" for first cousins.
}
\examples{
# Uncle and nephew plus an independent individual:
library(paramlink)
x <- nuclearPed(2)
x <- addSon(x, 4)
plot(x)
p <- c(.1, .2, .3, .4)
pNx <- pedNoA(x, c(3,6), p, 1)
plot(pNx)

# First order cousins with grandfather being homozygote
# with allele 3 and grandmother being heterozygotic with allele 1 and 2
y <- cousinsPed(1)
plot(y)
pNy <- pedNoA(y, c(5,8), p, marker_list = list(1, c(3,3), 2, c(1,2)))
plot(pNy)

}
\seealso{
\code{\link{paramlink}}, \code{\link{plot.pedNoA_}}
}
\author{
Mads Lindskou, \email{mads@math.aau.dk}
}
