% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA.R
\name{plot.pedNoA}
\alias{plot.pedNoA}
\title{A function for plotting the pedigree used in \code{pedNoA} along with information about which members and independent individuals are being tested.}
\usage{
\method{plot}{pedNoA}(x, title = "", ...)
}
\arguments{
\item{x}{A pedNoA object from \code{\link{pedNoA}}.}

\item{title}{The title of the plot}

\item{...}{Further arguments passed on to \code{plotPedList}}
}
\description{
Individuals in \code{x} which are being tested are grayed out in the plot.
}
\examples{
# Uncle and nephew plus an independent individual:
library(paramlink)
x <- nuclearPed(2)
x <- addSon(x, 4)
plot(x)
p <- c(.1, .2, .3, .4)
pNx <-pedNoA(x, c(3,6), list(p), 1)
plot(pNx)

# First order cousins with grandfather being homozygote
# with allele 3 and grandmother being heterozygotic with allele 1 and 2
y <- cousinsPed(1)
plot(y)
pNy <- pedNoA(y, c(5,8), list(p), marker_list = list(list(1), list(1, c(3,3), 2, c(1,2))))
plot(pNy)

}
\seealso{
\code{\link{pedNoA}}, \code{\link{pedNoA_}}
}
\author{
Mads Lindskou, \email{mads@math.aau.dk}
}
