% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA_.R
\name{plot.pedNoA_}
\alias{plot.pedNoA_}
\title{A function for plotting the pedigree used in \code{pedNoA_}}
\usage{
\method{plot}{pedNoA_}(x, title = "", ...)
}
\arguments{
\item{x}{A \code{pedNoA_} object}

\item{title}{The title of the plot}

\item{...}{Further arguments passed on to \code{plotPedList}}
}
\description{
Individuals in \code{x} which are being tested are grayed out in the plot.
}
\examples{
# 5 sibslings:
library(paramlink)
pNx <- pedNoA_("5sibs", p)
plot(pNx)

# Probability of being het or hom:
pN_1 <- pedNoA_("1sibs", p)
plot(pN_1)

# Uncle and nephew
pNy <- pedNoA_("uncle_nephew", p)
plot(pNy)

}
\seealso{
\code{\link{pedNoA}}, \code{\link{pedNoA_}}
}
\author{
Mads Lindskou, \email{mads@math.aau.dk}
}
