% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA.R
\name{print.pedNoA}
\alias{print.pedNoA}
\title{Printing a \code{pedNoA} object}
\usage{
\method{print}{pedNoA}(x, ...)
}
\arguments{
\item{x}{A pedNoA object from \code{\link{pedNoA}}.}

\item{...}{Further arguments passed on to \code{print.default}}
}
\description{
Printing information about the pedigree used in \code{pedNoA}
}
\examples{
# Uncle and nephew plus an independent individual:
library(paramlink)
x <- nuclearPed(2)
x <- addSon(x, 4)
plot(x)
p <- c(.1, .2, .3, .4)
pNx <-pedNoA(x, c(3,6), list(p), 1)
print(pNx)

# First order cousins with grandfather being homozygote
# with allele 3 and grandmother being heterozygotic with allele 1 and 2
y <- cousinsPed(1)
plot(y)
pNy <- pedNoA(y, c(5,8), list(p), marker_list = list(list(1), list(1, c(3,3), 2, c(1,2))))
print(pNy)

}
\seealso{
\code{\link{pedNoA}}, \code{\link{pedNoA_}}
}
\author{
Mads Lindskou, \email{mads@math.aau.dk}
}
