% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedNoA_.R
\name{print.pedNoA_}
\alias{print.pedNoA_}
\title{Printing a \code{pedNoA_} object}
\usage{
\method{print}{pedNoA_}(x, ...)
}
\arguments{
\item{x}{A pedNoA object from \code{\link{pedNoA}}}

\item{...}{Further arguments passed on to \code{print.default}}
}
\description{
Printing information about the pedigree used in \code{pedNoA_}
}
\examples{
# 5 sibslings:
library(paramlink)
pNx <-pedNoA_("5sibs", p)
plot(pNx)

# Probability of being het or hom:
pN_1 <- pedNoA_("1sibs", p)
plot(pN_1)

# Uncle and nephew
pNy <- pedNoA_("uncle_nephew", p)
plot(pNy)

}
\seealso{
\code{\link{pedNoA}}, \code{\link{pedNoA_}}
}
\author{
Mads Lindskou, \email{mads@math.aau.dk}
}
