% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pednoa.R
\name{pednoa}
\alias{pednoa}
\title{Pedigree Number of different Alleles}
\usage{
pednoa(
  afreq_list,
  mvec,
  founder_vec,
  nonfounder_list,
  evidence_list = NULL,
  locus_wise = TRUE,
  ncores = 1
)
}
\arguments{
\item{afreq_list}{A list with allele frequencies. One for each DNA loci}

\item{mvec}{Integer vector with the index of the pedigree members of interest}

\item{founder_vec}{Integer vector with the index of all founder}

\item{nonfounder_list}{List of three-dimensional index vectors of all
nonfounders on the form \code{(child, paternal_index, maternal_index)}}

\item{evidence_list}{A list of evidence. See the examples for details.}

\item{locus_wise}{Logical. \code{FALSE} if the pmfs should be convoluted}

\item{ncores}{Integer. The number of cores to be used in connection with
parallelizing some of the internal code.}
}
\description{
Find the probability distribution of m different pedigree members having
different number of alleles
}
\details{
There is no check whether the evidence is in agreement with
the pedigree. The user is responsible for entering correct evidence.
}
\examples{

# Allele frequencies:
# -------------------

p1 <- rbeta(6, 3, 6)
p1 <- p1 / sum(p1)

p2 <- rbeta(6, 3, 6)
p2 <- p2 / sum(p2)

# Evidence:
# -------------------

e <- list(
  list(3, c(1, 2), list(c(2, 2), c(1, 3))),
  list(1, 2, list(c(2, 3)))
)

# Individual 3 has genotype (2,2) on loci 1 and (1,3) on loci two.
# Individual 1 has genotype (2,2) on loci 2


# Consider the pedigree:
# ----------------------

#  1-------2
#      |
#  ---------
#  |       |
#  3       4-------6
#              |
#              5

# We let 1 and 5 be the individuals of interest. That is, we calculate the
# probability of seeing different alleles over two locis for 1 and 5 given
# the evidence.

p <- pednoa(
  list(p1, p2),
  c(1, 5),
  c(1, 2, 6),
  list(c(3, 1, 2), c(4, 1, 2), c(5, 4, 6)),
  e,
  locus_wise = FALSE
)

p

}
