% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pednoa.R
\name{pednoa}
\alias{pednoa}
\title{Pedigree Number of different Alleles}
\usage{
pednoa(afreq_list, mvec, founder_vec, nonfounder_list, evidence_list = NULL)
}
\arguments{
\item{afreq_list}{A list with allele frequencies. One for each DNA loci}

\item{mvec}{Integer vector with the index of the pedigree members of interest}

\item{founder_vec}{Integer vector with the index of all founder}

\item{nonfounder_list}{List of three-dimensional index vectors of all
nonfounders on the form \code{(child, paternal_index, maternal_index)}}

\item{evidence_list}{A list of evidence. See the examples for details.}
}
\description{
Find the probability distribution of m different pedigree members having
different number of alleles
}
\details{
There is no check whether the evidence is in agreement with
the pedigree. The user is responsible for entering correct evidence.
}
\examples{

# Allele frequencies:
# -------------------

p1 <- rbeta(6, 3,6)
p1 <- p1 / sum(p1)

p2 <- rbeta(6, 3,6)
p2 <- p2 / sum(p2)

p3 <- rbeta(6, 3,6)
p3 <- p3 / sum(p3)

p4 <- rbeta(6, 3, 6)
p4 <- p4 / sum(p4)

# Evidence:
# -------------------

e <- list(
  list(3, c(1, 2), list(c(2, 2), c(1, 3))),
  list(1, c(2, 3, 4), list(c(2, 3), c(1, 1), c(2, 3)))
)

# Individual 3 has genotype (2,2) on loci 1 and (1,1) on loci two.
# Individual 1 has genotype (2,3) on loci 2, (1,1) on loci 3 and (2,3) on loci 4 


# Consider the pedigree:
# ----------------------

#  1-------2
#      |
#  ---------
#  |       |
#  3       4-------6
#              |
#              5

# We let 1 and 5 be the individuals of interest. That is, we calculate the
# probability of seeing 4,3,...,16 different alleles over four locis for 1 and 5
# Note: They cannot have 16 different alleles due to the evidence.

p <- pednoa(
  list(p1, p2, p3, p4),
  c(1, 5),
  c(1, 2, 6),
  list(c(3, 1, 2), c(4, 1, 2), c(5, 4, 6)),
  e
)

}
